rename : Kiosk_Lv5 폴더로 이동
