+++ src/kiosk/Kiosk_Lv5/Kiosk.java
@@ -1,4 +1,4 @@
-package kiosk.Kiosk_Lv4;
+package kiosk.Kiosk_Lv5;
 
 import java.util.List;
 import java.util.Scanner;
